refactor: cross-domain dead code cleanup and diagnostic fixes
A comprehensive audit of all 8 generator domains identified 14
technical debt items. This PR addresses all of them.

Dead code removal (no breaking change):
- SignalR: remove unused ifaceSyntax parameter from TryAddMethod
  and TryAddProperty (Parser.cs)
- Nats: remove unused ifaceSyntax parameter from TryAddProperty
  + remove unused ParameterNames field from NatsMemberModel
  + remove dead names local from BuildParameters
- WebSocket: remove unused MessageName field from
  WebSocketMemberModel + remove dead GetMessageName and
  GetMessageNameFromProperty methods
- Mqtt: remove unused ParameterNames field from MqttMemberModel
  + remove dead names local from BuildParameters
- Grpc: remove redundant ternary in Emitter (both branches
  returned the same value)
- Events: remove unused context parameter from
  GenerateEventInterfacesSource (R3 + Reactive, 4 sites)
- Shared: remove unused IsPartial() extension method
  + delete unused ReactiveBackend enum from Observables.Core

Diagnostic message fixes:
- Mqtt OBS5005: "SystemReactive package required" →
  "Observables.Mqtt.Reactive package required"
- WebSocket OBS6005: "SystemReactive package required" →
  "Observables.WebSocket.Reactive package required"

Breaking change (public API, acceptable at 0.1.x):
- WebSocket: remove unused endpointName parameter and
  EndpointName property from WebSocketAttribute
  (PublicAPI.Unshipped.txt updated for 3 TFMs)

Code deduplication:
- Mqtt: extract duplicated TopicMatches method from
  MqttObservable.cs and SystemReactiveMqttAdapter.cs into
  new internal MqttTopicMatcher class with wildcard constants
  (InternalsVisibleTo added for Observables.Mqtt.Reactive)

Formatting:
- Grpc: normalize 5 csproj files (remove excessive blank lines)

Verified: Nuke Ci green (Clean/Restore/Compile/UnitTest, 2:40).
