+++ Observables.Grpc/Observables.Grpc.R3.SourceGenerators.Tests/Observables.Grpc.R3.SourceGenerators.Tests.csproj
@@ -1,51 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-
-
   <PropertyGroup>
-
     <RootNamespace>Observables.Grpc.R3.SourceGenerators.Tests</RootNamespace>
-
   </PropertyGroup>
 
-
-
   <ItemGroup>
-
     <Using Include="Xunit" />
-
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
-
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" />
-
     <PackageReference Include="R3" />
-
     <PackageReference Include="Verify.XunitV3" />
-
     <PackageReference Include="xunit.v3" />
-
     <PackageReference Include="xunit.runner.visualstudio">
-
       <PrivateAssets>all</PrivateAssets>
-
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-
     </PackageReference>
-
   </ItemGroup>
-
-
 
   <ItemGroup>
-
     <ProjectReference Include="..\Observables.Grpc\Observables.Grpc.csproj" />
-
     <ProjectReference Include="..\Observables.Grpc.R3.SourceGenerators\Observables.Grpc.R3.SourceGenerators.csproj" />
-
   </ItemGroup>
-
-
 
 </Project>
-
-
