+++ Observables.Grpc/Observables.Grpc.Reactive/Observables.Grpc.Reactive.csproj
@@ -12,4 +12,3 @@
   </ItemGroup>
 
 </Project>
-
