+++ Observables.Mqtt/Observables.Mqtt/MqttObservable.cs
@@ -44,7 +44,7 @@ public static class MqttObservable
         {
             async Task Handler(MqttApplicationMessageReceivedEventArgs e)
             {
-                if (!TopicMatches(topicFilter, e.ApplicationMessage.Topic))
+                if (!MqttTopicMatcher.Matches(topicFilter, e.ApplicationMessage.Topic))
                 {
                     return;
                 }
@@ -80,35 +80,4 @@ public static class MqttObservable
                 client.ApplicationMessageReceivedAsync -= Handler;
             }
         });
-
-    static bool TopicMatches(string filter, string? topic)
-    {
-        if (topic is null)
-        {
-            return false;
-        }
-
-        var filterParts = filter.Split('/');
-        var topicParts = topic.Split('/');
-        for (var i = 0; i < filterParts.Length; i++)
-        {
-            if (i >= topicParts.Length)
-            {
-                return false;
-            }
-
-            var fp = filterParts[i];
-            if (fp == "#")
-            {
-                return true;
-            }
-
-            if (fp != "+" && fp != topicParts[i])
-            {
-                return false;
-            }
-        }
-
-        return filterParts.Length == topicParts.Length;
-    }
 }
