+++ Observables.Nats/Observables.Nats.SourceGenerators.Shared/Parser.cs
@@ -96,8 +96,7 @@ internal static class Parser
                                 subscribeAttribute,
                                 observableType,
                                 members,
-                                diagnostics,
-                                ifaceSyntax);
+                                diagnostics);
                             break;
                     }
                 }
@@ -220,7 +219,7 @@ internal static class Parser
             return;
         }
 
-        var (declarations, names, hasCt) = BuildParameters(method);
+        var (declarations, hasCt) = BuildParameters(method);
         members.Add(
             new NatsMemberModel(
                 method.Name,
@@ -230,7 +229,6 @@ internal static class Parser
                 returnDisplay,
                 resultType,
                 declarations.ToImmutableEquatableArray(),
-                names.ToImmutableEquatableArray(),
                 subjectParameterNames.ToImmutableEquatableArray(),
                 hasCt,
                 payloadParameterName,
@@ -244,8 +242,7 @@ internal static class Parser
         INamedTypeSymbol? subscribeAttribute,
         INamedTypeSymbol? observableType,
         List<NatsMemberModel> members,
-        List<Diagnostic> diagnostics,
-        InterfaceDeclarationSyntax ifaceSyntax)
+        List<Diagnostic> diagnostics)
     {
         if (publishAttribute is not null && HasAttribute(property, publishAttribute))
         {
@@ -327,7 +324,6 @@ internal static class Parser
                 true,
                 returnDisplay,
                 resultType,
-                ImmutableEquatableArray.Empty<string>(),
                 ImmutableEquatableArray.Empty<string>(),
                 ImmutableEquatableArray.Empty<string>(),
                 false,
@@ -570,11 +566,10 @@ internal static class Parser
             out resultTypeDisplay,
             out returnTypeDisplay);
 
-    static (List<string> declarations, List<string> names, bool hasCancellationToken) BuildParameters(
+    static (List<string> declarations, bool hasCancellationToken) BuildParameters(
         IMethodSymbol method)
     {
         var declarations = new List<string>();
-        var names = new List<string>();
         var hasCt = false;
 
         for (var i = 0; i < method.Parameters.Length; i++)
@@ -588,11 +583,10 @@ internal static class Parser
                 continue;
             }
 
-            names.Add(parameter.Name);
             declarations.Add($"{parameter.Type.ToDisplayString(DisplayFormat)} {parameter.Name}");
         }
 
-        return (declarations, names, hasCt);
+        return (declarations, hasCt);
     }
 
     static bool IsCancellationToken(ITypeSymbol type) =>
