--- Observables.Shared/Observables.Core/ReactiveBackend.cs
@@ -1,10 +1,0 @@
-namespace Observables;
-
-/// <summary>
-/// Identifies which reactive runtime a generated API targets.
-/// </summary>
-public enum ReactiveBackend
-{
-    SystemReactive,
-    R3,
-}
