+++ Observables.WebSocket/Observables.WebSocket.SourceGenerators.Shared/Models/WebSocketMemberModel.cs
@@ -2,7 +2,6 @@ namespace Observables.WebSocket.Generators;
 
 internal sealed record WebSocketMemberModel(
     string MemberName,
-    string MessageName,
     WebSocketBoundaryKind BoundaryKind,
     bool IsProperty,
     string ReturnTypeDisplay,
