+++ Observables.WebSocket/Observables.WebSocket/Attributes.cs
@@ -2,10 +2,7 @@ namespace Observables.WebSocket;
 
 /// <summary>Marks a WebSocket proxy interface for source generation.</summary>
 [AttributeUsage(AttributeTargets.Interface)]
-public sealed class WebSocketAttribute(string? endpointName = null) : Attribute
-{
-    public string? EndpointName { get; } = endpointName;
-}
+public sealed class WebSocketAttribute : Attribute;
 
 /// <summary>Client-to-server message send mapped to WebSocket send.</summary>
 [AttributeUsage(AttributeTargets.Method)]
